+++ packages/react/test/visual/components/Arrow.tsx
@@ -1,14 +1,14 @@
 import {
-  Arrow,
   arrow,
   autoUpdate,
+  FloatingArrow,
   offset,
   Placement,
   useFloating,
 } from '@floating-ui/react';
 import {useRef, useState} from 'react';
 
-import type {Props} from '../../../src/components/Arrow';
+import type {Props} from '../../../src/components/FloatingArrow';
 
 const ROUND_D =
   'M0 20C0 20 2.06906 19.9829 5.91817 15.4092C7.49986 13.5236 8.97939 12.3809 10.0002 12.3809C11.0202 12.3809 12.481 13.6451 14.0814 15.5472C17.952 20.1437 20 20 20 20H0Z';
@@ -62,7 +62,7 @@ function Demo({
           }}
         >
           Tooltip
-          <Arrow
+          <FloatingArrow
             context={context}
             ref={arrowRef}
             staticOffset={edgeAlignment ? '15%' : null}
